requirements.txt
beautifulsoup4
request
pytest==3.9.1
